server/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic[email]==2.6.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP & WebSocket
aiohttp==3.9.1
httpx==0.24.1
websockets==12.0

# Data Analysis & Financial
pandas==2.1.1
numpy==1.26.0
scipy==1.11.3
yfinance==0.2.31
ta==0.11.0  # Technical Analysis
pandas-ta==0.3.14b  # Additional TA indicators

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.9.1
isort==5.12.0
mypy==1.5.1